pyproject.toml
[project]
name = "deeplearning-project-template"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0,<3.0.0",
    "transformers>=4.35.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0",
    "wandb>=0.16.0",
    "scikit-learn>=1.3.0",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "notebook>=7.4.7",
    "ruff>=0.14.2",
]
